refactor: 좋아요 등록 API 관련 리팩토링
