+++ src/network/api/index.ts
@@ -3,5 +3,6 @@ import * as userAPI from './user';
 import * as mypageAPI from './mypage';
 import * as qnaAPI from './qna';
 import * as tradeAPI from './trade';
+import * as postAPI from './post';
 
-export { rehomingAPI, userAPI, mypageAPI, qnaAPI, tradeAPI };
+export { rehomingAPI, userAPI, mypageAPI, qnaAPI, tradeAPI, postAPI };
