+++ src/network/api/trade.ts
@@ -43,17 +43,3 @@ export const postTrade: ApiHandler = (
     url: addPrefix(''),
     ...options,
   });
-
-// TODO: 공통 사용 여부 논의
-// 좋아요 등록
-export const postLikedStatus: ApiHandler = (
-  options, // postType, id
-) =>
-  instance({
-    headers: {
-      Authorization: `Bearer ${Token}`,
-    },
-    method: 'POST',
-    url: `${config.server.host}/api/v1/likes/${options.postType}/${options.id}`,
-    ...options,
-  });
